Optimized library structure with Python package best practices
