+++ dnet/datasets/binary_tiny_mnist/__init__.py
@@ -1,6 +1,6 @@
 from typing import Tuple
 
-from dnet.utils.tensor import Tensor
+from ...utils.tensor import Tensor
 
 
 def load_data() -> Tuple[Tensor, Tensor]:
