+++ dnet/datasets/tiny_mnist/__init__.py
@@ -1,4 +1,4 @@
-from dnet.utils.tensor import Tensor
+from ...utils.tensor import Tensor
 
 
 def load_data():
