+++ dnet/layers/__init__.py
@@ -1,1 +1,7 @@
-from jax.experimental.stax import Dense, Sigmoid, Relu
+from jax.experimental.stax import Dense, Relu, Sigmoid
+
+__all__ = [
+    "Dense",
+    "Relu",
+    "Sigmoid"
+]
