+++ dnet/optimizers/__init__.py
@@ -1,2 +1,8 @@
 from jax.experimental.optimizers import Optimizer, OptimizerState
 from jax.experimental.optimizers import sgd as SGD
+
+__all__ = [
+    "Optimizer",
+    "OptimizerState",
+    "SGD"
+]
