+++ dnet/utils/__init__.py
@@ -1,0 +1,13 @@
+from .interpreter import Interpreter
+from .tensor import Tensor, convert_to_tensor
+from .serialization import load_module, save_module
+from .trainer import Trainer
+
+__all__ = [
+    "Interpreter",
+    "Tensor",
+    "convert_to_tensor",
+    "load_module",
+    "save_module",
+    "Trainer"
+]
